PCHR-1868: Move Logic for interracting with Comments API's into LeaveRequestComment Service. Add tests. Remove unnecessary tests
